modificaciones referentes proceso flujo caja

+++ BDT-ejb/ejbModule/minsal/divap/vo/TransferenciaSummaryVO.java
@@ -9,25 +9,25 @@ public class TransferenciaSummaryVO implements Serializable{
 	 * 
 	 */
 	private static final long serialVersionUID = 4212552704753536630L;
-	private Integer porcentaje;
+	private Double porcentaje;
 	private Long monto;
 	
 	public TransferenciaSummaryVO() {
 		super();
-		this.porcentaje = 0;
+		this.porcentaje = 0.0;
 		this.monto = 0L;
 	}
 	
-	public TransferenciaSummaryVO(Integer porcentaje, Long monto) {
+	public TransferenciaSummaryVO(Double porcentaje, Long monto) {
 		super();
 		this.porcentaje = porcentaje;
 		this.monto = monto;
 	}
 	
-	public Integer getPorcentaje() {
+	public Double getPorcentaje() {
 		return porcentaje;
 	}
-	public void setPorcentaje(Integer porcentaje) {
+	public void setPorcentaje(Double porcentaje) {
 		this.porcentaje = porcentaje;
 	}
 	public Long getMonto() {
